Internationalization
This adds support for three UI languages (English, German, Hungarian),
as well as all FlaggedRevs wikis. To achieve the latter, it now uses the
page history instead of an abuse filter rule (which exists only on
huwiki). This makes the output longer: now all pages are reported that
have pending global renames, even if the user is trusted on wiki, but
their edit was not autoreviewed due to prior pending changes. (I think
it also makes more sense.)
The UI localization is not the best solution (separate templates for
each language, repeating the HTML code), but I don’t want to invest in
a better but more complicated solution unless non-techy people want to
participate in the translation.